[kitsune] Sanitize some types in the runtime
`__kit*_num_threads` was changed to return `uint64`. 32-bits is more than
enough for the actual, but in most uses of this, it will be immediately cast to
 64  bits anyway, so we might as well just return that from the outset.

 The thread launch functions, and thread functions themselves, were all changed
 to accept unsigned 64-bit integers instead of signed 64-bit integers. Once
 again, all uses for this is guaranteed to be unsigned because the start will
 always be 0, and the end will be at least 1.

 The corresponding `@llvm.kit.cpu.num.threads` intrinsic was also changed to
 return a 64-bit integer.

 This is now consistent with the `@llvm.kit.reduce.num.partials` intrinsic and
 the corresponding runtime functions that it is lowered to, which also return
 a 64-bit integer.

+++ kitsune/runtime/cuda/kitcuda.cpp
@@ -253,13 +253,13 @@ void __kitcuda_finalize() {
 /// The number of partial reductions to perform in parallel.
 ///
 /// \param n The trip count of the parallel loop in containing a reduction
-extern "C" int64_t __kitcuda_reduce_num_partials(int64_t n) {
+extern "C" uint64_t __kitcuda_reduce_num_partials(uint64_t n) {
   kitrt::log("kitcuda", "Calculating number of partial reductions\n");
 
   // FIXME: This is simply a placeholder to check that the rest of the
   // transformations work as expected. It is beyond terrible for performance, so
   // fix this is ASAP.
-  int numPartials = 8;
+  uint64_t numPartials = 8;
 
   kitrt::log("kitcuda", "Number of partial reductions: %ld\n", numPartials);
